[CORE] Use unsigned long instead of u32 for vmm_delay APIs
Currently, we use u32 parameters for vmm_delay APIs which means
even on 64bit host we will endup implementing arch_delay_xxxx()
using 32bit registers.

This patch updates all vmm_delay APIs to use unsiged long instead
of u32 so that we can use unsigned long for arch_delay_xxxx(). This
will help us use 64bit registers on 64bit host for implementing
arch_delay_xxxx() functions.

+++ arch/arm/cpu/arm32ve/include/arch_delay.h
@@ -26,11 +26,11 @@
 #include <vmm_types.h>
 
 /** Low-level delay loop */
-void arch_delay_loop(u32 count);
+void arch_delay_loop(unsigned long count);
 
 /** Estimated cycles for given loop count 
  *  Note: This can be processor specific
  */
-u32 arch_delay_loop_cycles(u32 count);
+unsigned long arch_delay_loop_cycles(unsigned long count);
 
 #endif
